# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notes API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/notebooks/": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebook"
                ],
                "summary": "Getting a list of user notebooks",
                "responses": {
                    "200": {
                        "description": "Notebooks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.GetAllNotebooksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebook"
                ],
                "summary": "Creating notebook",
                "parameters": [
                    {
                        "description": "Notebook data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreateUpdateNotebookDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notebook ID",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateNotebookResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/notebooks/{notebook-id}": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebook"
                ],
                "summary": "Updating user notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New notebook data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreateUpdateNotebookDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebook"
                ],
                "summary": "Deleting user notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/notebooks/{notebook-id}/notes/": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Getting a list of user notes in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.GetAllNotesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Creating a note in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Note data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note ID",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateNoteResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
        "/api/notebooks/{notebook-id}/notes/{note-id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Getting a note in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.GetNoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Deleting a note from a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Updating a note in a notebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notebook ID",
                        "name": "notebook-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "note-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New note data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PatchNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK status",
                        "schema": {
                            "$ref": "#/definitions/docs.OkStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "$ref": "#/definitions/docs.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.CreateNoteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "docs.CreateNotebookResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "docs.CreateUpdateNotebookDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "docs.GetAllNotebooksResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "notebooks": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "description": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "docs.GetAllNotesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "body": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "docs.GetNoteResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "docs.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "docs.OkStatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "domain.CreateNoteDTO": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.PatchNoteDTO": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}